Time: 22 ms (65.95%), Space: 22.5 MB (32.47%) - LeetHub

+++ 637-average-of-levels-in-binary-tree/637-average-of-levels-in-binary-tree.cpp
@@ -18,7 +18,7 @@ public:
         q.push(root);
         while(!q.empty())
         {
-           int n = q.size();
+           double n = q.size();
             double sum=0;
 //vector<int> level;
             for(int i=0; i<n; i++)
